Implementing new shadui buttons to gallery
