--- src/components/Button/Button.fixture.tsx
@@ -1,6 +1,0 @@
-import Button from './Button';
-
-export default {
-  Default: <Button>Default button</Button>,
-  Filled: <Button filled>Filled button</Button>,
-};
